requirements.txt
ipp-ds
holidays
tqdm
matplotlib
importlib-metadata
openpyxl
numpy<1.20.0
pandas
python-dotenv
setuptools
wheel
xlrd
tslearn
plotly
seaborn
scikit-learn
scipy
statsmodels
ipykernel
mlxtend
mlflow
gitpython
prophet
